feat: 新增获取网关数据hooks
