+++ src/dashboard-front-new/package.json
@@ -23,7 +23,7 @@
   "license": "ISC",
   "dependencies": {
     "art-template": "^4.13.2",
-    "bkui-vue": "0.0.1-beta.213",
+    "bkui-vue": "0.0.2-beta.60",
     "connect-history-api-fallback": "^2.0.0",
     "cookie": "^0.5.0",
     "cookie-parser": "^1.4.5",
